x3dna parsing util class
