+++ RNAFoldAssess/utils/__init__.py
@@ -1,4 +1,5 @@
 from .dssr import *
 from .pdb_tools import *
 from .normalizers import *
+from .secondary_structure_tools import *
 from .pipelines import *
